+++ CTestTestfile.cmake
@@ -1,0 +1,8 @@
+# CMake generated Testfile for 
+# Source directory: /home/tiger/CascadeProjects/oatpp-mariadb
+# Build directory: /home/tiger/CascadeProjects/oatpp-mariadb
+# 
+# This file includes the relevant testing commands required for 
+# testing this directory and lists subdirectories to be tested as well.
+subdirs("src")
+subdirs("test")