+++ test/CMakeFiles/oatpp-mariadb-tests.dir/DependInfo.cmake
@@ -1,0 +1,25 @@
+
+# Consider dependencies only in project.
+set(CMAKE_DEPENDS_IN_PROJECT_ONLY OFF)
+
+# The set of languages for which implicit dependencies are needed:
+set(CMAKE_DEPENDS_LANGUAGES
+  )
+
+# The set of dependency files which are needed:
+set(CMAKE_DEPENDS_DEPENDENCY_FILES
+  "/home/tiger/CascadeProjects/oatpp-mariadb/test/oatpp-mariadb/tests.cpp" "test/CMakeFiles/oatpp-mariadb-tests.dir/oatpp-mariadb/tests.cpp.o" "gcc" "test/CMakeFiles/oatpp-mariadb-tests.dir/oatpp-mariadb/tests.cpp.o.d"
+  "/home/tiger/CascadeProjects/oatpp-mariadb/test/oatpp-mariadb/types/TypeMappingTest.cpp" "test/CMakeFiles/oatpp-mariadb-tests.dir/oatpp-mariadb/types/TypeMappingTest.cpp.o" "gcc" "test/CMakeFiles/oatpp-mariadb-tests.dir/oatpp-mariadb/types/TypeMappingTest.cpp.o.d"
+  "" "test/oatpp-mariadb-tests" "gcc" "test/CMakeFiles/oatpp-mariadb-tests.dir/link.d"
+  )
+
+# Targets to which this target links which contain Fortran sources.
+set(CMAKE_Fortran_TARGET_LINKED_INFO_FILES
+  )
+
+# Targets to which this target links which contain Fortran sources.
+set(CMAKE_Fortran_TARGET_FORWARD_LINKED_INFO_FILES
+  )
+
+# Fortran module output directory.
+set(CMAKE_Fortran_TARGET_MODULE_DIR "")