+++ test/CMakeFiles/oatpp-mariadb-tests.dir/cmake_clean.cmake
@@ -1,0 +1,14 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/oatpp-mariadb-tests.dir/link.d"
+  "CMakeFiles/oatpp-mariadb-tests.dir/oatpp-mariadb/tests.cpp.o"
+  "CMakeFiles/oatpp-mariadb-tests.dir/oatpp-mariadb/tests.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb-tests.dir/oatpp-mariadb/types/TypeMappingTest.cpp.o"
+  "CMakeFiles/oatpp-mariadb-tests.dir/oatpp-mariadb/types/TypeMappingTest.cpp.o.d"
+  "oatpp-mariadb-tests"
+  "oatpp-mariadb-tests.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/oatpp-mariadb-tests.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()