+++ test/oatpp-mariadb/tests.cpp
@@ -19,26 +19,24 @@
 #include "oatpp-mariadb/types/AnyTypeTest.hpp"
 #include "oatpp-mariadb/migration/SchemaVersionTest.hpp"
 #include "oatpp-mariadb/QueryResultTest.hpp"
-
 #include "oatpp-mariadb/transaction/TransactionTest.hpp"
 #include "oatpp-mariadb/crud/CrudTest.hpp"
 #include "oatpp-mariadb/crud/ProductCrudTest.hpp"
 #include "oatpp-mariadb/crud/EnhancedCrudTest.hpp"
+#include "oatpp-mariadb/types/TypeMappingTest.hpp"
 
-#include "oatpp/core/base/Environment.hpp"
-
+#include "oatpp-test/UnitTest.hpp"
 #include <iostream>
+
 
 namespace {
 
 void runTests() {
-  OATPP_RUN_TEST(oatpp::test::mariadb::QueryResultTest);
-  /*
   OATPP_RUN_TEST(oatpp::test::mariadb::ql_template::ParserTest);
+  OATPP_RUN_TEST(oatpp::test::mariadb::types::NumericTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::Int64Test);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::Float64Test);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::UInt8Test);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::NumericTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::DateTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::DateTimeTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::VarCharTest);
@@ -47,30 +45,25 @@ void runTests() {
   OATPP_RUN_TEST(oatpp::test::mariadb::types::TimeTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::YearTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::EnumTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::transaction::TransactionTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::StringTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::TimeTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::YearTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::EnumTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::SetTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::JsonTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::types::BinaryTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::TimeTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::YearTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::EnumTest);
+  OATPP_RUN_TEST(oatpp::test::mariadb::types::Int32Test);
+  OATPP_RUN_TEST(oatpp::test::mariadb::types::StringTest);
+  OATPP_RUN_TEST(oatpp::test::mariadb::types::AnyTypeTest);
+  OATPP_RUN_TEST(oatpp::test::mariadb::migration::SchemaVersionTest);
+  OATPP_RUN_TEST(oatpp::test::mariadb::QueryResultTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::transaction::TransactionTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::crud::CrudTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::crud::ProductCrudTest);
   OATPP_RUN_TEST(oatpp::test::mariadb::crud::EnhancedCrudTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::types::AnyTypeTest);
-  OATPP_RUN_TEST(oatpp::test::mariadb::migration::SchemaVersionTest);
-  */
+  OATPP_RUN_TEST(oatpp::test::mariadb::types::TypeMappingTest);
 }
 
 }
 
 int main() {
-
+  
   oatpp::base::Environment::init();
 
   runTests();